requirements.txt
genanki
pandas
pywsd
requests
colorama
eng-to-ipa
python-dotenv
openai